Refactor main screen
