+++ src/components/CreateTarget/styles.js
@@ -1,0 +1,11 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  newTarget: {
+    height: 100,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default styles;
