+++ src/utils/helpers.js
@@ -1,4 +1,6 @@
+import { Animated } from 'react-native';
 import queryString from 'query-string';
+import { ANIMATIONS_DURATION } from 'constants/common';
 
 export const applyQueryParams = (url, params) => {
   const queryParams = queryString.stringify(params);
@@ -7,4 +9,11 @@ export const applyQueryParams = (url, params) => {
 
 export const getStringWithCondition = (condition, first, second) => {
   return condition ? first : second;
+};
+
+export const animate = (animatedValue, toValue) => {
+  Animated.timing(animatedValue, {
+    toValue,
+    duration: ANIMATIONS_DURATION,
+  }).start();
 };
